Reference UPM packages
Summary:
Removes Oculus plugin from Assets, import them via UPM instead. The experience is buggy in subtle ways, but can still be completed. The process to create this diff was as follows:

1. Delete the individual plugin folders from Assets/Oculus. You may need to close your project and do this in the Explorer/Finder.
2. Open the project. There will be errors, but "Ignore" them and open the broken project (NOT in Safe Mode)
3. Import the plugins you deleted, but from the Asset store so they appear in your manifest: {F1754650053}
4. The compile errors should go away, but you may need to manually correct some things if there are remaining errors.

A complete fix for the experience bugs will come next, after migrating the project to MRUK.

NOTE: this diff also removes the Sandcastle build file. This is deemed acceptable because: the app has been removed from App Lab so we don't need any more apks, and we also don't want to modify the prior apks because QA likes to use TWB to test Scene backwards-compatibility.

Reviewed By: yagizmunganmeta, yoavgoldsteinfb

Differential Revision: D52040901

fbshipit-source-id: ef422408d931160827e96f3d7ac3dc87c66535c3